requirements.txt
streamlit==1.40.2
pandas==2.2.3
plotly==5.24.1
anthropic==0.40.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
reportlab==4.2.5
Pillow==10.4.0
openpyxl==3.1.5